# backend-node/openapi.yaml
openapi: 3.0.3
info:
  title: AI Services Platform API
  version: 2.0.0
  description: APIs for AI-powered services including Virtual Assistant, IDP, Computer Vision, Authentication, and more.
  contact:
    name: AI Services Platform Support
    email: support@aiservicesplatform.com
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.yourdomain.com
    description: Production server
...
paths:
...
  /api/assistant-channels:
    get:
      summary: Get all assistant channel configurations for tenant
      tags: [Assistant Channels]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of channel configurations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  channels:
                    type: array
                    items:
                      $ref: '#/components/schemas/AssistantChannels'
        '401':
          description: Unauthorized
    patch:
      summary: Update assistant channel configuration
      tags: [Assistant Channels]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                voice:
                  type: object
                chat:
                  type: object
              required: [productId]
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  channels:
                    $ref: '#/components/schemas/AssistantChannels'
        '401':
          description: Unauthorized
...
  /api/assistant-channels/by-phone/{phoneNumber}:
    get:
      summary: Get assistant channels by phone number
      tags: [Assistant Channels]
      description: Public endpoint for telephony providers (no auth required)
      parameters:
        - in: path
          name: phoneNumber
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Channel configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantChannels'
        '404':
          description: Phone number not found
  /api/assistant-channels/{productId}:
    get:
      summary: Get assistant channel configuration for a product
      tags: [Assistant Channels]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Channel configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantChannels'
        '404':
          description: Channel not found
        '401':
          description: Unauthorized

  # Prompts Endpoints
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie for authentication
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AssistantChannels:
      type: object
      properties:
        _id:
          type: string
        tenantId:
          type: string
        productId:
          type: string
        voice:
          $ref: '#/components/schemas/VoiceChannel'
        chat:
          $ref: '#/components/schemas/ChatChannel'
        sms:
          $ref: '#/components/schemas/SmsChannel'
        whatsapp:
          $ref: '#/components/schemas/WhatsAppChannel'
    VoiceChannel:
      type: object
      properties:
        enabled:
          type: boolean
        phoneNumber:
          type: string
        fallbackNumber:
          type: string
        promptTemplateId:
          type: string
        customPrompts:
          type: object
        ragConfig:
          type: object
        promptContext:
          type: object
        voiceSettings:
          type: object
          properties:
            language:
              type: string
            voiceId:
              type: string
    ChatChannel:
      type: object
      properties:
        enabled:
          type: boolean
        promptTemplateId:
          type: string
        customPrompts:
          type: object
        ragConfig:
          type: object
        promptContext:
          type: object
        greeting:
          type: string
        typingIndicator:
          type: boolean
        maxTurns:
          type: integer
    SmsChannel:
      type: object
      properties:
        enabled:
          type: boolean
        phoneNumber:
          type: string
    WhatsAppChannel:
      type: object
      properties:
        enabled:
          type: boolean
        businessAccountId:
          type: string
...